+++ libs/designer-ui/src/lib/expressioneditor/index.tsx
@@ -2,7 +2,7 @@ import { MonacoEditor as Editor, EditorLanguage } from '../editor/monaco';
 import type { EventHandler } from '../eventhandler';
 import type { editor } from 'monaco-editor';
 import type { MutableRefObject } from 'react';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
 export interface ExpressionEditorEvent {
   value: string;
@@ -13,14 +13,34 @@ export interface ExpressionEditorEvent {
 export interface ExpressionEditorProps {
   initialValue: string;
   editorRef?: MutableRefObject<editor.IStandaloneCodeEditor | null>;
+  isDragging: boolean;
+  dragDistance?: number;
+  currentHeight: number;
+  setCurrentHeight: (height: number) => void;
   onBlur?: EventHandler<ExpressionEditorEvent>;
+  setIsDragging: (isDragging: boolean) => void;
+  setExpressionEditorError: (error: string) => void;
 }
 
-export function ExpressionEditor({ initialValue, editorRef, onBlur }: ExpressionEditorProps): JSX.Element {
-  const [focused, setFocused] = useState(false);
+export function ExpressionEditor({
+  initialValue,
+  editorRef,
+  onBlur,
+  isDragging,
+  dragDistance,
+  currentHeight,
+  setCurrentHeight,
+  setIsDragging,
+  setExpressionEditorError,
+}: ExpressionEditorProps): JSX.Element {
+  const [mouseDownLocation, setMouseDownLocation] = useState(0);
+  useEffect(() => {
+    if (isDragging && dragDistance) {
+      setCurrentHeight(Math.min(Math.max(100, 100 + dragDistance - mouseDownLocation), 200));
+    }
+  }, [isDragging, dragDistance, mouseDownLocation, currentHeight, setCurrentHeight]);
 
   const handleBlur = (): void => {
-    setFocused(false);
     if (onBlur && editorRef?.current) {
       const currentSelection = editorRef.current.getSelection();
       const currentCursorPosition = editorRef.current.getPosition()?.column ?? 1 - 1;
@@ -34,11 +54,8 @@ export function ExpressionEditor({ initialValue, editorRef, onBlur }: Expression
     }
   };
 
-  const handleFocus = (): void => {
-    setFocused(true);
-  };
-
   const handleChangeEvent = (e: editor.IModelContentChangedEvent): void => {
+    setExpressionEditorError('');
     const changedText = e.changes.length ? e.changes[0].text : '';
     if (changedText === '\r\n' && editorRef?.current) {
       const oldPosition = editorRef.current.getPosition();
@@ -69,12 +86,10 @@ export function ExpressionEditor({ initialValue, editorRef, onBlur }: Expression
   };
 
   return (
-    <div className={focused ? 'msla-expression-editor-container msla-focused' : 'msla-expression-editor-container'}>
+    <div className="msla-expression-editor-container" style={{ height: currentHeight }}>
       <Editor
         ref={editorRef}
-        className={'msla-expression-editor-main'}
         language={EditorLanguage.templateExpressionLanguage}
-        folding={false}
         lineNumbers="off"
         value={initialValue}
         scrollbar={{ horizontal: 'hidden', vertical: 'hidden' }}
@@ -82,11 +97,22 @@ export function ExpressionEditor({ initialValue, editorRef, onBlur }: Expression
         overviewRulerLanes={0}
         overviewRulerBorder={false}
         contextMenu={false}
-        onFocus={handleFocus}
         onBlur={handleBlur}
         onContentChanged={handleChangeEvent}
-        width={'340px'}
+        width={'100%'}
+        wordWrap="bounded"
+        wordWrapColumn={200}
+        automaticLayout={true}
       />
+      <div
+        className="msla-expression-editor-expand"
+        onMouseDown={(e) => {
+          setMouseDownLocation(e.clientY);
+          setIsDragging(true);
+        }}
+      >
+        <div className="msla-expression-editor-expand-icon" /> <div className="msla-expression-editor-expand-icon-2" />
+      </div>
     </div>
   );
 }