+++ libs/designer-ui/src/lib/tokenpicker/tokenpickersection/tokenpickersection.tsx
@@ -1,7 +1,7 @@
+import { getWindowDimensions, TokenPickerMode } from '..';
 import type { ValueSegment } from '../../editor';
 import type { ExpressionEditorEvent } from '../../expressioneditor';
 import type { TokenGroup } from '../models/token';
-import { TokenPickerMode } from '../tokenpickerpivot';
 import { TokenPickerNoDynamicContent } from './tokenpickernodynamiccontent';
 import { TokenPickerNoMatches } from './tokenpickernomatches';
 import type { GetValueSegmentHandler } from './tokenpickeroption';
@@ -17,8 +17,9 @@ interface TokenPickerSectionProps {
   searchQuery: string;
   expressionEditorRef: MutableRefObject<editor.IStandaloneCodeEditor | null>;
   expression: ExpressionEditorEvent;
-  editMode: boolean;
-  isDynamicContentAvailable: boolean;
+  fullScreen: boolean;
+  noDynamicContent: boolean;
+  expressionEditorCurrentHeight: number;
   setExpression: Dispatch<SetStateAction<ExpressionEditorEvent>>;
   getValueSegmentFromToken: GetValueSegmentHandler;
   tokenClickedCallback?: (token: ValueSegment) => void;
@@ -30,25 +31,47 @@ export const TokenPickerSection = ({
   searchQuery,
   expressionEditorRef,
   expression,
-  editMode,
-  isDynamicContentAvailable,
+  fullScreen,
+  noDynamicContent,
+  expressionEditorCurrentHeight,
   setExpression,
   getValueSegmentFromToken,
   tokenClickedCallback,
 }: TokenPickerSectionProps): JSX.Element => {
-  const [tokenLength, setTokenLength] = useState(new Array<number>(tokenGroup.length));
+  const [dynamicTokenLength, setDynamicTokenLength] = useState(new Array<number>(tokenGroup.length));
+  const [expressionTokenLength, setExpressionTokenLength] = useState(new Array<number>(expressionGroup.length));
   const [noItems, setNoItems] = useState(false);
 
   useEffect(() => {
-    setNoItems(tokenLength.reduce((sum, a) => sum + a, 0) === 0);
-  }, [searchQuery, tokenLength]);
+    if (selectedKey === TokenPickerMode.TOKEN_EXPRESSION || selectedKey === TokenPickerMode.TOKEN) {
+      setNoItems(dynamicTokenLength.reduce((sum, a) => sum + a, 0) === 0);
+    } else {
+      setNoItems(expressionTokenLength.reduce((sum, a) => sum + a, 0) === 0);
+    }
+  }, [dynamicTokenLength, expressionTokenLength, searchQuery, selectedKey]);
+
+  const [windowDimensions, setWindowDimensions] = useState(getWindowDimensions());
+
+  useEffect(() => {
+    function handleResize() {
+      setWindowDimensions(getWindowDimensions());
+    }
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
 
   return (
-    <div className="msla-token-picker-sections">
-      {isDynamicContentAvailable || selectedKey === TokenPickerMode.EXPRESSION ? (
-        <>
-          {searchQuery ? <TokenPickerNoMatches noItems={noItems} /> : null}
-          {(selectedKey === TokenPickerMode.TOKEN ? tokenGroup : expressionGroup).map((section, i) => {
+    <div
+      className="msla-token-picker-sections"
+      style={{ maxHeight: fullScreen ? windowDimensions.height - (expressionEditorCurrentHeight + 287) : 550 }}
+    >
+      {searchQuery && noItems ? <TokenPickerNoMatches /> : null}
+      {noDynamicContent && (selectedKey === TokenPickerMode.TOKEN_EXPRESSION || selectedKey === TokenPickerMode.TOKEN) ? (
+        <TokenPickerNoDynamicContent />
+      ) : (
+        (selectedKey === TokenPickerMode.TOKEN_EXPRESSION || selectedKey === TokenPickerMode.TOKEN ? tokenGroup : expressionGroup).map(
+          (section, i) => {
             if (section.tokens.length > 0) {
               return (
                 <div key={`token-picker-section-${i}`} className={'msla-token-picker-section'}>
@@ -57,8 +80,7 @@ export const TokenPickerSection = ({
                     section={section}
                     searchQuery={searchQuery}
                     index={i}
-                    setTokenLength={setTokenLength}
-                    editMode={editMode}
+                    setTokenLength={selectedKey === TokenPickerMode.EXPRESSION ? setExpressionTokenLength : setDynamicTokenLength}
                     expressionEditorRef={expressionEditorRef}
                     expression={expression}
                     setExpression={setExpression}
@@ -69,10 +91,8 @@ export const TokenPickerSection = ({
               );
             }
             return null;
-          })}
-        </>
-      ) : (
-        <TokenPickerNoDynamicContent />
+          }
+        )
       )}
     </div>
   );